iter1 commit 25, add env vars processing
